+++ examples/test_libmeshb_block.c
@@ -1,10 +1,10 @@
 
-// libMeshb 7.5 basic example: read a quad mesh, split it into triangles
+// libMeshb 8.0 basic example: read a quad mesh, split it into triangles
 // and write the result back using fast block transfer
 
 #include <stdio.h>
 #include <stdlib.h>
-#include <libmeshb7.h>
+#include <libmeshb8.h>
 
 #ifdef PRINTF_INT64_MODIFIER
 #define INT64_T_FMT "%" PRINTF_INT64_MODIFIER "d"