+++ examples/test_libmeshb_block02.f90
@@ -1,4 +1,4 @@
-! libMeshb 7.79 example: transform a quadrilateral mesh into a triangular one
+! libMeshb 8.0 example: transform a quadrilateral mesh into a triangular one
 ! using fast block transfer
 
 ! test_libmeshb_block02_f90
@@ -11,7 +11,7 @@
 program test_libmeshb_block02_f90
   !>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
   use iso_fortran_env
-  use libmeshb7
+  use libmeshb8
   !<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<
   !>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
   implicit none
@@ -222,6 +222,7 @@ program test_libmeshb_block02_f90
   deallocate(VerTab,VerRef)
   deallocate(QadTab,QadRef)
   deallocate(TriTab,TriRef)
+  deallocate(fields,fieldNames)
   deallocate(solTab)
   !<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<