+++ examples/test_polyhedra.c
@@ -3,8 +3,8 @@
 
 #include <stdio.h>
 #include <stdlib.h>
-#include <libmeshb7.h>
-#include <libmeshb7_helpers.h>
+#include <libmeshb8.h>
+#include <libmeshb8_helpers.h>
 
 int main()
 {