+++ sources/libmeshb8.h
@@ -2,22 +2,22 @@
 
 /*----------------------------------------------------------------------------*/
 /*                                                                            */
-/*                               LIBMESHB V7.84                               */
+/*                               LIBMESHB V8.02                               */
 /*                                                                            */
 /*----------------------------------------------------------------------------*/
 /*                                                                            */
 /*    Description:        handle .meshb file format I/O                       */
 /*    Author:             Loic MARECHAL                                       */
 /*    Creation date:      dec 09 1999                                         */
-/*    Last modification:  oct 09 2024                                         */
+/*    Last modification:  mar 12 2026                                         */
 /*                                                                            */
 /*----------------------------------------------------------------------------*/
 
 // [Bruno] get PRINTF_INT64_MODIFIER
 // #include <geogram/third_party/pstdint.h>
 
-#ifndef LIBMESHB7_H
-#define LIBMESHB7_H
+#ifndef LIBMESHB8_H
+#define LIBMESHB8_H
 
 #include <stdint.h>
 
@@ -26,29 +26,31 @@
 /* Defines                                                                    */
 /*----------------------------------------------------------------------------*/
 
-#define GmfStrSiz       1024
-#define FilStrSiz       64
-#define GmfMaxTyp       1000
-#define GmfMaxKwd       GmfLastKeyword - 1
-#define GmfMshVer       1
-#define GmfRead         1
-#define GmfWrite        2
-#define GmfSca          1
-#define GmfVec          2
-#define GmfSymMat       3
-#define GmfMat          4
-#define GmfFloat        8
-#define GmfDouble       9
-#define GmfInt          10
-#define GmfLong         11
-#define GmfFloatVec     12
-#define GmfDoubleVec    13
-#define GmfIntTab       14
-#define GmfIntVec       14
-#define GmfLongTab      15
-#define GmfLongVec      15
-#define GmfArgTab       100
-#define GmfArgLst       101
+#define GmfStrSiz             1024
+#define FilStrSiz             64
+#define GmfMaxTyp             1000
+#define GmfMaxKwd             GmfLastKeyword - 1
+#define GmfMshVer             1
+#define GmfRead               1
+#define GmfWrite              2
+#define GmfSca                1
+#define GmfVec                2
+#define GmfSymMat             3
+#define GmfMat                4
+#define GmfFloat              8
+#define GmfDouble             9
+#define GmfInt                10
+#define GmfLong               11
+#define GmfFloatVec           12
+#define GmfDoubleVec          13
+#define GmfIntTab             14
+#define GmfIntVec             14
+#define GmfLongTab            15
+#define GmfLongVec            15
+#define GmfStartParallelWrite 16
+#define GmfStopParallelWrite  17
+#define GmfArgTab             100
+#define GmfArgLst             101
 
 enum GmfKwdCod
 {
@@ -305,20 +307,23 @@ extern "C" {
 /* Public procedures                                                          */
 /*----------------------------------------------------------------------------*/
 
-extern int64_t GmfOpenMesh          (const char *, int, ...);
-extern int     GmfCloseMesh         (int64_t);
-extern int64_t GmfStatKwd           (int64_t, int, ...);
-extern int     GmfSetKwd            (int64_t, int, int64_t, ...);
-extern int     GmfGotoKwd           (int64_t, int);
-extern int     GmfGetLin            (int64_t, int, ...);
-extern int     GmfSetLin            (int64_t, int, ...);
-extern int     GmfGetBlock          (int64_t, int, int64_t, int64_t, int, void *, void *, ...);
-extern int     GmfSetBlock          (int64_t, int, int64_t, int64_t, int, void *, void *, ...);
-extern char   *GmfReadByteFlow      (int64_t, int *);
-extern int     GmfWriteByteFlow     (int64_t, char *, int);
-extern int     GmfSetHONodesOrdering(int64_t, int, int *, int *);
-extern int     GmfGetFloatPrecision (int64_t);
-extern void    GmfSetFloatPrecision (int64_t, int);
+extern int64_t GmfOpenMesh             (const char *, int, ...);
+extern int     GmfCloseMesh            (int64_t);
+extern int     GmfCloseUnfinishedMesh  (int64_t);
+extern int64_t GmfStatKwd              (int64_t, int, ...);
+extern int     GmfSetKwd               (int64_t, int, int64_t, ...);
+extern int     GmfGotoKwd              (int64_t, int);
+extern int     GmfGetLin               (int64_t, int, ...);
+extern int     GmfSetLin               (int64_t, int, ...);
+extern int     GmfGetBlock             (int64_t, int, int64_t, int64_t,
+                                        int, void *, void *, ...);
+extern int     GmfSetBlock             (int64_t, int, int64_t, int64_t,
+                                        int, void *, void *, ...);
+extern char   *GmfReadByteFlow         (int64_t, int *);
+extern int     GmfWriteByteFlow        (int64_t, char *, int);
+extern int     GmfSetHONodesOrdering   (int64_t, int, int *, int *);
+extern int     GmfGetFloatPrecision    (int64_t);
+extern void    GmfSetFloatPrecision    (int64_t, int);
 
 
 /*----------------------------------------------------------------------------*/